+++ src/Pages/Likes/likes.jsx
@@ -1,5 +1,5 @@
 import {useUserData} from '../../Context/userDataContext'
-import DynamicVideoCard from '../../Components/Video Card/dynamic-videCard'
+import DynamicVideoCard from '../../Components/Video Card/dynamic-videoCard'
 import '../../Utilities/CSS/Utilities.css'
 
 function Likes() {
